1.3.0
fixed bugs and errror messages within the console. Fixed the nav bar links so that when they are clicked you are directed to the correct potion of the page and so that when the PORTFOLIO is clicked you are returned to the top of the screen right now this doesnt make much sense since the navbar is not sticky. updated form so that styles match the style of the page. added in a form element and the inputs when clicked to be logged to the send button is clicked.